Optimize Day24 by removing linear search
removed redundant mapping that removed blizzard direction information in each iteration over new position evaluation. Removed blizzard direction as precomptation
Mapping over a set gives a list as a result, and so doing a search on the result will be a linear search O(n) and not a hashed search o(1).
Reduced runtime from 80secs to 0.4secs